Move `_replaceIf` so `twisted.python.util` can use `_oldStyle`.
