+++ src/twisted/python/util.py
@@ -14,8 +14,6 @@ try:
 except ImportError:
     setgroups = getgroups = None
 
-from functools import wraps
-
 from twisted.python.compat import _PY3, unicode
 from incremental import Version
 from twisted.python.deprecate import deprecatedModuleAttribute
@@ -993,37 +991,6 @@ def runWithWarningsSuppressed(suppressedWarnings, f, *args, **kwargs):
         for a, kw in suppressedWarnings:
             warnings.filterwarnings(*a, **kw)
         return f(*args, **kwargs)
-
-
-
-def _replaceIf(condition, alternative):
-    """
-    If C{condition}, replace this function with C{alternative}.
-
-    @param condition: A L{bool} which says whether this should be replaced.
-
-    @param alternative: An alternative function that will be swapped in instead
-        of the original, if C{condition} is truthy.
-
-    @return: A decorator.
-    """
-    def decorator(func):
-
-        if condition is True:
-            call = alternative
-        elif condition is False:
-            call = func
-        else:
-            raise ValueError(("condition argument to _replaceIf requires a "
-                              "bool, not {}").format(repr(condition)))
-
-        @wraps(func)
-        def wrapped(*args, **kwargs):
-            return call(*args, **kwargs)
-
-        return wrapped
-
-    return decorator
